refactor: Move SiteAdminPage to pages directory and remove HealthBridge Enhanced from navigation - Update import paths for SiteAdminPage and clean up navigation configuration by removing the HealthBridge Enhanced entry. Update lastUpdated timestamp in content cache for consistency.
