+++ src/config/navigation.ts
@@ -83,12 +83,6 @@ export const protectedProjectItems = [
     requiresAuth: true,
   },
   {
-    title: "HealthBridge Enhanced",
-    url: "protected/healthbridge-enhanced",
-    icon: Target,
-    requiresAuth: true,
-  },
-  {
     title: "Settings",
     url: "protected/settings",
     icon: Settings,
