+++ src/data/content-cache.json
@@ -2557,7 +2557,7 @@
     "portfolioCount": 14,
     "blogCount": 10,
     "projectCount": 1,
-    "lastUpdated": "2025-09-03T01:18:10.443Z",
+    "lastUpdated": "2025-09-03T01:51:27.652Z",
     "version": "1.0.0"
   }
 }
