+++ src/pages/SiteAdminPage.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { Button } from './ui/button';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from './ui/card';
-import { P } from './ui/typography';
-import { Shield, ArrowRight, Loader2, UserCheck, Briefcase, CheckCircle, XCircle, Activity, Database, Globe, BarChart3, Settings, Users, Mail, Code, FileText, Target } from 'lucide-react';
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { P } from '@/components/ui/typography';
+import { Shield, ArrowRight, Loader2, UserCheck, Briefcase, CheckCircle, XCircle, Activity, Database, Globe, BarChart3, Settings, Users, Mail, Code, FileText } from 'lucide-react';
 import { useAuth } from '../hooks/useAuth';
 import { testAIWorker } from '../api/contactAnalyzer';
 import { cachedContentService } from '@/api/cachedContentService';
@@ -556,17 +556,6 @@ export const SiteAdminPage: React.FC = () => {
                       <div className="flex items-center space-x-2">
                         <FileText className="h-4 w-4 text-teal-600" />
                         <span className="font-medium text-teal-800">Content Studio</span>
-                      </div>
-                      <ArrowRight className="h-4 w-4 text-teal-600" />
-                    </a>
-                    
-                    <a
-                      href="/protected/healthbridge-enhanced"
-                      className="flex items-center justify-between p-3 bg-white rounded-lg border border-teal-200 hover:bg-teal-50 transition-colors duration-200"
-                    >
-                      <div className="flex items-center space-x-2">
-                        <Target className="h-4 w-4 text-teal-600" />
-                        <span className="font-medium text-teal-800">HealthBridge Enhanced</span>
                       </div>
                       <ArrowRight className="h-4 w-4 text-teal-600" />
                     </a>
